+++ src/pyclarify/fields/dataframe.py
@@ -14,34 +14,26 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from pydantic import BaseModel, constr, validate_arguments, Extra, validator
-from pydantic.fields import Optional
-from typing import List, Union, Dict
-from typing_extensions import Literal
-from datetime import datetime, timedelta
-from enum import Enum
-from pydantic.json import timedelta_isoformat
+# TODO: Check if we can outsource the dataframe
+
+from .constraints import InputID, NumericalValuesType
+from datetime import datetime
+from pydantic import BaseModel, constr, conint, validate_arguments, validator
+from typing import List, Dict
 from pyclarify.__utils__.auxiliary import local_import
 
-# constrained string defined by the API
-InputID = constr(regex=r"^[a-z0-9_-]{1,40}$")
-ResourceID = constr(regex=r"^[a-v0-9]{20}$")
-LabelsKey = constr(regex=r"^[A-Za-z0-9-_/]{1,40}$")
-AnnotationKey = constr(regex=r"^[A-Za-z0-9-_/]{1,40}$")
-NumericalValuesType = List[Union[float, int, None]]
-SHA1Hash = constr(regex=r"^[0-9a-f]{5,40}$")
 
-class DataFrame(BaseModel):
+class DataFrame(BaseModel):  # <- Views
     times: List[datetime] = None
     series: Dict[InputID, NumericalValuesType] = None
 
-    @validator('series')
+    @validator("series", allow_reuse=True)
     def convert_numpy_to_native(cls, v):
         if isinstance(v, Dict):
             for key, value in v.items():
                 v[key] = [None if x != x else x for x in value]
         return v
-        
+
     def to_pandas(self):
         """Convert the instance into a pandas DataFrame.
 
@@ -115,76 +107,3 @@ def merge(dataframes: List[DataFrame]):
         series[signal] = values
 
     return DataFrame(times=times, series=series)
-
-
-class DataQuery(BaseModel, extra=Extra.forbid):
-    include: bool = False
-    notBefore: Optional[datetime]
-    before: Optional[datetime]
-    rollup: Union[timedelta, Literal["window"]] = None
-
-
-class ResourceQuery(BaseModel, extra=Extra.forbid):
-    include: bool = False
-    filter: dict  # TODO: ResourceFilter (https://docs.clarify.io/v1.1/reference/filtering)
-    limit: int = (
-        0  # select_items: max=50, default=10 | select_signal: max=1000, default=50
-    )
-    skip: int = 0
-
-
-class GenericSummary(BaseModel, extra=Extra.forbid):
-    id: ResourceID
-    created: bool
-
-
-class InsertSummary(GenericSummary, extra=Extra.forbid):
-    pass
-
-
-class SaveSummary(GenericSummary, extra=Extra.forbid):
-    updated: bool
-
-
-class TypeSignal(str, Enum):
-    numeric = "numeric"
-    enum = "enum"
-
-
-class SourceTypeSignal(str, Enum):
-    measurement = "measurement"
-    aggregation = "aggregation"
-    prediction = "prediction"
-
-
-class SignalInfo(BaseModel):
-    name: str
-    type: TypeSignal = TypeSignal.numeric
-    description: str = ""
-    labels: Dict[LabelsKey, List[str]] = {}
-    annotations: Optional[Dict[AnnotationKey, str]] = {}
-    engUnit: str = ""
-    enumValues: Dict[str, str] = {}
-    sourceType: SourceTypeSignal = SourceTypeSignal.measurement
-    sampleInterval: timedelta = None
-    gapDetection: timedelta = None
-
-    class Config:
-        json_encoders = {timedelta: timedelta_isoformat}
-        extra = Extra.forbid
-
-
-class ResourceMetadata(BaseModel):
-    contentHash: str 
-    updatedAt: datetime
-    createdAt: datetime
-
-
-class Signal(SignalInfo):
-    item: Union[ResourceID, None]
-    inputId: InputID
-    meta: ResourceMetadata
-
-
-class Item(SignalInfo):
-    visible: bool = False
