+++ src/pyclarify/query/__init__.py
@@ -14,4 +14,16 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from .filter import Filter, Equal, NotEqual, Regex, In, NotIn, LessThan, GreaterThan, GreaterThanOrEqual, Operators, Comparison
+from .filter import (
+    Filter,
+    Equal,
+    NotEqual,
+    Regex,
+    In,
+    NotIn,
+    LessThan,
+    GreaterThan,
+    GreaterThanOrEqual,
+    Operators,
+    Comparison,
+)
