+++ src/pyclarify/views/resource.py
@@ -13,9 +13,3 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-
-from pyclarify.client import ClarifyClient
-from pyclarify.views.signals import Signal, SignalInfo
-from pyclarify.views.items import Item, ItemInfo
-from pyclarify.fields.dataframe import DataFrame
-import pyclarify.query
