+++ tests/test_models_response.py
@@ -5,8 +5,17 @@ from pydantic.error_wrappers import ValidationError
 
 
 sys.path.insert(1, "src/")
-from pyclarify.models.response import *
-
+from pyclarify.fields.error import Error, ErrorData
+from pyclarify.views.generics import Response
+from pyclarify.views.dataframe import InsertResponse
+from pyclarify.views.items import (
+    SelectItemsResponse,
+    PublishSignalsResponse,
+)
+from pyclarify.views.signals import (
+    SelectSignalsResponse,
+    SaveSignalsResponse,
+)
 
 class TestError(unittest.TestCase):
     # TODO: Make more thorough
